feat: Implement login and logout functionality
The code changes include the implementation of login and logout functionality. This allows users to log in with their email and password, and log out from the app.
